Interpolate price data only (#4)
Price data from coingecko may be missing when merged with transaction
data from the blockchain, so it needs to be interpolated to fill in the
missing data. However pandas does not like to interpolate certain
datatypes (like datetime64ns). When the timestamp was added to the
dataframe in 446f6ca, this seems to have broken the interpolation
function. So instead, this change only interpolates the price data.